+++ src/math/matrix44.h
@@ -21,10 +21,10 @@
 class alignas(32) Matrix44
 {
     /// access to modifiable element by index
-    real& operator[](size_t i)       { return val[i]; }
+    real& operator[](unsigned i)       { return val[i]; }
     
     /// access element value by index
-    real  operator[](size_t i) const { return val[i]; }
+    real  operator[](unsigned i) const { return val[i]; }
 
 public:
 
@@ -36,7 +36,7 @@ public:
     /// copy constructor
     Matrix44(Matrix44 const& M)
     {
-        for ( unsigned u = 0;  u < 16; ++u )
+        for ( unsigned u = 0; u < 16; ++u )
             val[u] = M.val[u];
     }
     
@@ -96,14 +96,14 @@ public:
     /// set all elements to zero
     void reset()
     {
-        for ( unsigned u = 0;  u < 16; ++u )
+        for ( unsigned u = 0; u < 16; ++u )
             val[u] = 0.0;
     }
     
     /// true if any value is different from 'zero'
     bool operator != (real zero) const
     {
-        for ( unsigned u = 0;  u < 16; ++u )
+        for ( unsigned u = 0; u < 16; ++u )
             if ( val[u] != zero )
                 return true;
         return false;
@@ -114,20 +114,20 @@ public:
 
     /// conversion to array of 'real'
     real* data() { return val; }
-    real* addr(const size_t i, const size_t j) { return val + ( i + 4*j ); }
+    real* addr(const unsigned i, const unsigned j) { return val + ( i + 4*j ); }
 
     /// access functions to element by line and column indices
-    real& operator()(const size_t i, const size_t j)       { return val[i+4*j]; }
-    real  operator()(const size_t i, const size_t j) const { return val[i+4*j]; }
+    real& operator()(const unsigned i, const unsigned j)       { return val[i+4*j]; }
+    real  operator()(const unsigned i, const unsigned j) const { return val[i+4*j]; }
     
     /// extract column vector at given index
-    Vector4 column(const size_t i) const
+    Vector4 column(const unsigned i) const
     {
         return Vector4(val+4*i);
     }
     
     /// extract line vector at given index
-    Vector4 line(const size_t i) const
+    Vector4 line(const unsigned i) const
     {
         return Vector4(val[i], val[4+i], val[8+i], val[12+i]);
     }
@@ -159,9 +159,9 @@ public:
         std::ostringstream os;
         os.precision(p);
         os << "[";
-        for ( unsigned i = 0;  i < 4; ++i )
+        for ( unsigned i = 0; i < 4; ++i )
         {
-            for ( unsigned j = 0;  j < 4; ++j )
+            for ( unsigned j = 0; j < 4; ++j )
                 os << " " << std::fixed << std::setw(w) << (*this)(i,j);
             if ( i < 2 )
                 os << ";";
@@ -174,7 +174,7 @@ public:
     /// scale all elements
     void scale(const real alpha)
     {
-        for ( unsigned u = 0;  u < 16; ++u )
+        for ( unsigned u = 0; u < 16; ++u )
             val[u] *= alpha;
     }
     
@@ -188,7 +188,7 @@ public:
     const Matrix44 operator -() const
     {
         Matrix44 M;
-        for ( unsigned u = 0;  u < 16; ++u )
+        for ( unsigned u = 0; u < 16; ++u )
             M.val[u] = -val[u];
         return M;
     }
@@ -197,7 +197,7 @@ public:
     const Matrix44 operator *(const real alpha) const
     {
         Matrix44 M;
-        for ( unsigned u = 0;  u < 16; ++u )
+        for ( unsigned u = 0; u < 16; ++u )
             M.val[u] = val[u] * alpha;
         return M;
     }
@@ -212,7 +212,7 @@ public:
     const Matrix44 operator +(Matrix44 const& M) const
     {
         Matrix44 res;
-        for ( unsigned u = 0;  u < 16; ++u )
+        for ( unsigned u = 0; u < 16; ++u )
             res.val[u] = val[u] + M.val[u];
         return res;
     }
@@ -226,7 +226,7 @@ public:
         store4(val+8, add4(load4(val+8), load4(M.val+8)));
         store4(val+12, add4(load4(val+12), load4(M.val+12)));
 #else
-        for ( unsigned u = 0;  u < 16; ++u )
+        for ( unsigned u = 0; u < 16; ++u )
             val[u] += M.val[u];
 #endif
     }
@@ -240,7 +240,7 @@ public:
         store4(val+8, sub4(load4(val+8), load4(M.val+8)));
         store4(val+12, sub4(load4(val+12), load4(M.val+12)));
 #else
-        for ( unsigned u = 0;  u < 16; ++u )
+        for ( unsigned u = 0; u < 16; ++u )
             val[u] -= M.val[u];
 #endif
     }
@@ -274,8 +274,8 @@ public:
         store4(res.val+8 , permute2f128(t0, t2, 0x31));
         store4(res.val+12, permute2f128(t1, t3, 0x31));
 #else
-        for ( unsigned x = 0;  x < 4; ++x )
-        for ( unsigned y = 0;  y < 4; ++y )
+        for ( unsigned x = 0; x < 4; ++x )
+        for ( unsigned y = 0; y < 4; ++y )
             res.val[y+4*x] = val[x+4*y];
 #endif
         return res;
@@ -451,7 +451,7 @@ public:
     void add_full(Matrix44 const& M)
     {
         real const* src = M.val;
-        for ( unsigned u = 0;  u < 16; ++u )
+        for ( unsigned u = 0; u < 16; ++u )
             val[u] += src[u];
     }
     
@@ -459,7 +459,7 @@ public:
     void add_full(const real alpha, Matrix44 const& M)
     {
         real const* src = M.val;
-        for ( unsigned u = 0;  u < 16; ++u )
+        for ( unsigned u = 0; u < 16; ++u )
             val[u] += alpha * src[u];
     }
     
@@ -467,7 +467,7 @@ public:
     void sub_full(Matrix44 const& M)
     {
         real const* src = M.val;
-        for ( unsigned u = 0;  u < 16; ++u )
+        for ( unsigned u = 0; u < 16; ++u )
             val[u] -= src[u];
     }
 
@@ -476,10 +476,10 @@ public:
     {
         real const* src = M.val;
 #if ( 1 )
-        for ( unsigned u = 0;  u < 16; ++u )
+        for ( unsigned u = 0; u < 16; ++u )
             val[u] += src[u];
 #else
-        for ( unsigned x = 0;  x < 4; ++x )
+        for ( unsigned x = 0; x < 4; ++x )
         for ( unsigned y = x; y < 4; ++y )
             val[y+4*x] += src[y+4*x];
 #endif
@@ -490,10 +490,10 @@ public:
     {
         real const* src = M.val;
 #if ( 1 )
-        for ( unsigned u = 0;  u < 16; ++u )
+        for ( unsigned u = 0; u < 16; ++u )
             val[u] += alpha * src[u];
 #else
-        for ( unsigned x = 0;  x < 4; ++x )
+        for ( unsigned x = 0; x < 4; ++x )
         for ( unsigned y = x; y < 4; ++y )
             val[y+4*x] += alpha * src[y+4*x];
 #endif
@@ -522,10 +522,10 @@ public:
     {
         real const* src = M.val;
 #if ( 1 )
-        for ( unsigned u = 0;  u < 16; ++u )
+        for ( unsigned u = 0; u < 16; ++u )
             val[u] -= src[u];
 #else
-        for ( unsigned x = 0;  x < 4; ++x )
+        for ( unsigned x = 0; x < 4; ++x )
         for ( unsigned y = x; y < 4; ++y )
             val[y+4*x] -= src[y+4*x];
 #endif
@@ -535,15 +535,15 @@ public:
     /// add all elements of block 'S' to array 'M'
     void addto(real * M, size_t ldd) const
     {
-        for ( unsigned x = 0;  x < 4; ++x )
-        for ( unsigned y = 0;  y < 4; ++y )
+        for ( unsigned x = 0; x < 4; ++x )
+        for ( unsigned y = 0; y < 4; ++y )
             M[y+ldd*x] = val[y+4*x];
     }
     
     /// add lower elements of this block to upper triangle of 'M'
     void addto_upper(real * M, size_t ldd) const
     {
-        for ( unsigned x = 0;  x < 4; ++x )
+        for ( unsigned x = 0; x < 4; ++x )
         for ( unsigned y = x; y < 4; ++y )
             M[y+ldd*x] = val[y+4*x];
     }
@@ -551,15 +551,15 @@ public:
     /// add all elements of this block to 'M', with transposition
     void addto_trans(real * M, size_t ldd) const
     {
-        for ( unsigned x = 0;  x < 4; ++x )
-        for ( unsigned y = 0;  y < 4; ++y )
+        for ( unsigned x = 0; x < 4; ++x )
+        for ( unsigned y = 0; y < 4; ++y )
             M[x+ldd*y] = val[y+4*x];
     }
     
     /// add lower elements of this block to both upper and lower triangles of 'M'
     void addto_symm(real * M, size_t ldd) const
     {
-        for ( unsigned x = 0;  x < 4; ++x )
+        for ( unsigned x = 0; x < 4; ++x )
         {
             M[x+ldd*x] = val[x+4*x];
             for ( unsigned y = x+1; y < 4; ++y )
@@ -683,9 +683,9 @@ inline std::ostream& operator << (std::ostream& os, Matrix44 const& mat)
     int w = (int)os.width();
     os.width(1);
     os << "[";
-    for ( unsigned i = 0;  i < 4; ++i )
+    for ( unsigned i = 0; i < 4; ++i )
     {
-        for ( unsigned j = 0;  j < 4; ++j )
+        for ( unsigned j = 0; j < 4; ++j )
             os << " " << std::fixed << std::setw(w) << mat(i,j);
         if ( i < 2 )
             os << ";";
